1/perce_mur_backend_new/requirements.txt
asgiref==3.9.1
boto3==1.40.32
botocore==1.40.32
Django==5.2.6
django-cors-headers==4.8.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
